requirements.txt
python-dotenv
certifi